// src/generated/model/programLevelResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProgramLevelResponseStatus } from './programLevelResponseStatus.ts';

export interface ProgramLevelResponse {
  id?: string;
  programId?: string;
  code?: string;
  name?: string;
  status?: ProgramLevelResponseStatus;
  displayOrder?: number;
  createdAt?: string;
  updatedAt?: string;
}